Enhancement: Support wheels
Wheels were previously unsupported becuase the version was incorrectly
computed and dynamodb key conflicts would occur if a wheel and tarball
were computed to have the same package name and version. This is likely to
occur so the package name is now the same as the filename. The api uses
the normalized name for lookups and thus the package name is only used
when handling events from S3.